backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
alembic==1.13.2
yfinance==0.2.40
requests==2.32.3
beautifulsoup4==4.12.3
APScheduler==3.10.4
python-dotenv==1.0.1
